refactoring project structure: websocket classes
